Revert "✨ feat: 공유 페이지 일기 피드 링크 설정 및 BackButton 뒤로가기 기능 구현 "
