--- grass-diary/.prettierrc.json
@@ -1,12 +1,0 @@
-{
-  "printWidth": 80,
-  "tabWidth": 2,
-  "useTabs": false,
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "endOfLine": "auto",
-  "bracketSpacing": true,
-  "arrowParens": "avoid",
-  "proseWrap": "never"
-}
