+++ grass-diary/package.json
@@ -12,9 +12,6 @@
     "@stylexjs/stylex": "^0.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.0",
-    "react-slick": "^0.30.1",
-    "slick-carousel": "^1.8.1",
     "vite-plugin-stylex": "^0.5.0"
   },
   "devDependencies": {
