--- grass-diary/share.md
@@ -1,1 +1,0 @@
-share branch
