--- grass-diary/src/components/Button.jsx
@@ -1,35 +1,0 @@
-import stylex from '@stylexjs/stylex';
-
-const styles = stylex.create({
-  startButton: (width, color, backgroundColor, border) => ({
-    marginTop: '25px',
-    padding: '12px 10px 12px 10px',
-
-    borderRadius: '30px',
-
-    fontWeight: '500',
-    cursor: 'pointer',
-
-    width,
-    color,
-    border,
-    backgroundColor,
-  }),
-});
-
-const Button = ({ text, onClick, width, color, backgroundColor, border }) => {
-  return (
-    <>
-      <button
-        onClick={onClick}
-        {...stylex.props(
-          styles.startButton(width, color, backgroundColor, border),
-        )}
-      >
-        {text}
-      </button>
-    </>
-  );
-};
-
-export default Button;
